fix: (Timer+Server) fixed Timer dllexport warnings + Boost Windows SDK version warnings + various other warnings
see:
https://stackoverflow.com/a/3001279
https://learn.microsoft.com/en-us/cpp/porting/modifying-winver-and-win32-winnt?view=msvc-170
https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-1-c4251?view=msvc-170
https://learn.microsoft.com/en-us/answers/questions/1088370/(c-)-(vs-2022)-how-to-ignore-warnings-from-externa

+++ Server/Server/Game.cpp
@@ -84,7 +84,8 @@ void Game::NextRound()
 		auto							   count = isWordsEntity ? m_db.count<WordsEntity>() : m_db.count<QuotesEntity>();
 		std::uniform_int_distribution<int> distribution(0, count - 1);*/
 
-		srand(time(0));
+        // remove srand
+		srand(static_cast<unsigned int>(time(0)));
 
 		int randomRoundIndex = rand();
